Correct integer array to array of reals for the modifier values.

+++ singularity-eos/eos/singularity_eos.f90
@@ -319,7 +319,7 @@ contains
     
     ! arrays for if optionals aren't included
     integer(kind=c_int), dimension(2) :: zero_mods
-    integer(kind=8), dimension(2)     :: zero_values
+    real(kind=8), dimension(2)     :: zero_values
     ! use arrays if present, else use 0s
     if (present(sg_mods_enabled) .and. present(sg_mods_values)) then
        err = init_sg_IdealGas(matindex-1, eos%ptr, gm1, Cv, &
@@ -349,7 +349,7 @@ contains
     
     ! arrays for if optionals aren't included
     integer(kind=c_int), dimension(2) :: zero_mods
-    integer(kind=8), dimension(2)     :: zero_values
+    real(kind=8), dimension(2)     :: zero_values
     ! use arrays if present, else use 0s
     if (present(sg_mods_enabled) .and. present(sg_mods_values)) then
        err = init_sg_Gruneisen(matindex-1, eos%ptr, C0, s1, s2, s3, G0, b, &
@@ -380,7 +380,7 @@ contains
     
     ! arrays for if optionals aren't included
     integer(kind=c_int), dimension(2) :: zero_mods
-    integer(kind=8), dimension(2)     :: zero_values
+    real(kind=8), dimension(2)     :: zero_values
     ! use arrays if present, else use 0s
     if (present(sg_mods_enabled) .and. present(sg_mods_values)) then
        err = init_sg_JWL(matindex-1, eos%ptr, A, B, R1, R2, w, rho0, Cv, &
@@ -410,7 +410,7 @@ contains
     
     ! arrays for if optionals aren't included
     integer(kind=c_int), dimension(2) :: zero_mods
-    integer(kind=8), dimension(2)     :: zero_values
+    real(kind=8), dimension(2)     :: zero_values
     ! use arrays if present, else use 0s
     if (present(sg_mods_enabled) .and. present(sg_mods_values)) then
        err = init_sg_DavisProducts(matindex-1, eos%ptr, a, b, k, n, vc, pc, &
@@ -443,7 +443,7 @@ contains
     
     ! arrays for if optionals aren't included
     integer(kind=c_int), dimension(2) :: zero_mods
-    integer(kind=8), dimension(2)     :: zero_values
+    real(kind=8), dimension(2)     :: zero_values
     ! use arrays if present, else use 0s
     if (present(sg_mods_enabled) .and. present(sg_mods_values)) then
        err = init_sg_DavisReactants(matindex-1, eos%ptr, rho0, e0, P0, T0, A, &
@@ -478,7 +478,7 @@ contains
     
     ! arrays for if optionals aren't included
     integer(kind=c_int), dimension(2) :: zero_mods
-    integer(kind=8), dimension(2)     :: zero_values
+    real(kind=8), dimension(2)     :: zero_values
     ! use arrays if present, else use 0s
     if (present(sg_mods_enabled) .and. present(sg_mods_values)) then
        err = init_sg_SpinerDependsRhoT(matindex-1, eos%ptr,&
@@ -512,7 +512,7 @@ contains
     
     ! arrays for if optionals aren't included
     integer(kind=c_int), dimension(2) :: zero_mods
-    integer(kind=8), dimension(2)     :: zero_values
+    real(kind=8), dimension(2)     :: zero_values
     ! use arrays if present, else use 0s
     if (present(sg_mods_enabled) .and. present(sg_mods_values)) then
        err = init_sg_SpinerDependsRhoSie(matindex-1, eos%ptr,&
@@ -544,7 +544,7 @@ contains
     
     ! arrays for if optionals aren't included
     integer(kind=c_int), dimension(2) :: zero_mods
-    integer(kind=8), dimension(2)     :: zero_values
+    real(kind=8), dimension(2)     :: zero_values
     ! use arrays if present, else use 0s
     if (present(sg_mods_enabled) .and. present(sg_mods_values)) then
        err = init_sg_eospac(matindex-1, eos%ptr, id, c_loc(sg_mods_enabled), &
